+++ nupic/research/temporal_memory_shim.py
@@ -29,7 +29,7 @@ import numpy
 from nupic.math import GetNTAReal
 from nupic.research.TP import TP
 from nupic.research.TP10X2 import TP10X2
-from nupic.research.temporal_memory import Connections
+from nupic.research.connections import Connections
 
 
 
@@ -76,7 +76,7 @@ class TemporalMemoryShim(TPClass):
       activationThreshold=activationThreshold,
       seed=seed)
 
-    self.connections = Connections(columnDimensions, cellsPerColumn)
+    self.connections = Connections(numberOfCols * cellsPerColumn)
     self.predictiveCells = set()
